requirements.txt
streamlit>=1.35.0
groq>=0.9.0
sentence-transformers>=3.0.0
numpy>=1.24.0
pyyaml>=6.0
